LayoutTests/http/wpt/web-animations/interfaces/Document/getAnimations-expected.txt


PASS Test document.getAnimations for non-animated content 
FAIL Test document.getAnimations for script-generated animations assert_equals: getAnimation only returns running animations expected 0 but got 2
FAIL Test the order of document.getAnimations with script generated animations assert_array_equals: getAnimations() returns running animations lengths differ, expected 2 got 4
FAIL Test document.getAnimations with null target Can't find variable: KeyframeEffectReadOnly
